requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.30.0
peft>=0.6.0
bitsandbytes>=0.40.0
accelerate>=0.20.0

# API
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utils
pyyaml>=6.0
tqdm>=4.65.0
psutil>=5.9.0

# Evaluation
nltk>=3.8.0
rouge-score>=0.1.2
